Create citrix-workspace.rb
Sorry I'm not sure how brew works, so I'm not sure I've filled in the details here correctly. 

Citrix Workspace was deleted from homebrew in https://github.com/Homebrew/homebrew-cask/pull/56881 but as far as I can tell it's now available from a static URL.

I've not tested the uninstall and zapping part of this specification (it's a copy-paste from the deleted file) so I might be causing a net frustration for the maintainers, close the PR if this isn't the proper way to go about things.